+++ components/vf-banner/CHANGELOG.md
@@ -1,3 +1,7 @@
+### 2.0.4
+
+* Changed: Test cases added for Angular  [Tracking issue](https://github.com/visual-framework/vf-core/issues/1932)
+
 ### 2.0.3
 
 * Dependency update [Tracking issue](https://github.com/visual-framework/vf-core/issues/1647)